Use i32 everywhere for counters
We actually should use u32 internally, but:
because Kotlin doesn't (yet) have unsigned integers, it's always going to pass us signed values.
We also do all the error checking on the Rust side, so we need to receive the signed value to check for negativity and bail out.

We still don't do overflow checks though.

+++ glean-core/android/src/main/java/mozilla/telemetry/glean/private/CounterMetricType.kt
@@ -64,7 +64,7 @@ class CounterMetricType(
             LibGleanFFI.INSTANCE.glean_counter_add(
                 Glean.handle,
                 this@CounterMetricType.handle,
-                amount.toLong())
+                amount)
         }
     }
 
@@ -96,10 +96,9 @@ class CounterMetricType(
      */
     @VisibleForTesting(otherwise = VisibleForTesting.NONE)
     fun testGetValue(pingName: String = sendInPings.first()): Int {
-        // FIXME(#19): glean-core should give us an int to begin with
         if (!testHasValue(pingName)) {
             throw NullPointerException()
         }
-        return LibGleanFFI.INSTANCE.glean_counter_test_get_value(Glean.handle, this.handle, pingName).toInt()
+        return LibGleanFFI.INSTANCE.glean_counter_test_get_value(Glean.handle, this.handle, pingName)
     }
 }
